+++ src/main/java/edu/netcracker/project/logistic/model/Person.java
@@ -1,7 +1,5 @@
 package edu.netcracker.project.logistic.model;
 
-import org.springframework.format.annotation.DateTimeFormat;
-
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.time.LocalDateTime;
