+++ src/main/resources/templates/admin/admin_crud_employee.html
@@ -57,22 +57,6 @@
                             </div>
                         </div>
 
-                        <div class="form-group required">
-                            <label for="password" class="control-label">Password</label>
-                            <input type="password" class="form-control" id="password" placeholder=""
-                                   required="required" minlength="6" maxlength="30"
-                                   th:value="${password}"
-                                   th:field="*{password}" />
-                        </div>
-
-                        <div class="form-group required">
-                            <label for="repeat_password" class="control-label">Repeat password</label>
-                            <input type="password" class="form-control" id="repeat_password" placeholder=""
-                                   required="required" minlength="6" maxlength="30"
-                                   oninput="validate(this)"
-                                   th:value="${password}" />
-                        </div>
-
                         <div class="form-group required"
                              th:classappend="${#fields.hasErrors('contact.email')}? 'has-error'">
                             <label for="email" class="control-label">E-mail</label>
